+++ libs/core/src/lib/validations/empty/empty.ts
@@ -1,6 +1,6 @@
 import { LengthProperty } from '../../types/properties';
 import { SyncValidation } from '../../types/types';
-import { createValidation } from '../create-validation-fn';
+import { createValidation } from '../create-validation';
 
 /**
  * Creates a validation function that checks if the values length is 0.
