+++ libs/core/src/lib/validations/equals/equals.ts
@@ -1,5 +1,5 @@
 import { SyncValidation } from '../../types/types';
-import { createValidation } from '../create-validation-fn';
+import { createValidation } from '../create-validation';
 import { DEFAULT_PLACEHOLDERS } from '../message-formatter';
 
 /**
