+++ libs/core/src/lib/validations/greater-than-or-equals/greater-than-or-equals.ts
@@ -1,6 +1,6 @@
 import { NumberProperty } from '../../types/properties';
 import { SyncValidation } from '../../types/types';
-import { createValidation } from '../create-validation-fn';
+import { createValidation } from '../create-validation';
 import { DEFAULT_PLACEHOLDERS } from '../message-formatter';
 
 /**
