+++ libs/core/src/lib/validations/is-truthy/is-truthy.ts
@@ -1,5 +1,5 @@
 import { SyncValidation } from '../../types/types';
-import { createValidation } from '../create-validation-fn';
+import { createValidation } from '../create-validation';
 
 /**
  * Creates a validation function that checks if the value is truthy.
