+++ libs/core/src/lib/validations/less-than-or-equals/less-than-or-equals.ts
@@ -1,6 +1,6 @@
 import { NumberProperty } from '../../types/properties';
 import { SyncValidation } from '../../types/types';
-import { createValidation } from '../create-validation-fn';
+import { createValidation } from '../create-validation';
 import { DEFAULT_PLACEHOLDERS } from '../message-formatter';
 
 /**
