+++ libs/core/src/lib/validations/must/must.ts
@@ -1,5 +1,5 @@
 import { SyncValidation } from '../../types/types';
-import { createValidation } from '../create-validation-fn';
+import { createValidation } from '../create-validation';
 
 /**
  * Creates a validation function that checks if the value meets the specified criteria.
