+++ libs/core/src/lib/validations/not/not.ts
@@ -1,5 +1,5 @@
 import { SyncValidation } from '../../types/types';
-import { createValidation } from '../create-validation-fn';
+import { createValidation } from '../create-validation';
 
 /**
  * Creates an inverted validation function.
